requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
configparser==5.2.0
Django==4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.0.0
e==1.4.5
greenlet==1.1.2
idna==3.3
numpy==1.22.0
pandas==1.3.5
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tzdata==2021.5
unicodecsv==0.14.1
urllib3==1.26.8
